Refactor: 로그인이 필요한 페이지는 (afterLogin) 폴더 아래로 이동
