+++ app/(afterLogin)/introduceTrainer/components/PictureExist.tsx
