+++ app/(afterLogin)/schedule/components/ScheduleComponent.tsx
