+++ app/(afterLogin)/schedule/style/customCalendar.css
