+++ app/(afterLogin)/trainerMyPage/layout.tsx
@@ -1,0 +1,19 @@
+import { ReactNode } from "react";
+import BottomNavigationBar from "@/components/BottomNavigationBar";
+import { SmallHeader } from "@/components/SmallHeader";
+
+export default async function trainerMyPageLayout({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  return (
+    <div className=" md:max-w-md h-screen">
+      <SmallHeader title="마이페이지" arrrowHidden />
+
+      <div>{children}</div>
+
+      <BottomNavigationBar />
+    </div>
+  );
+}
